Fix IOException being thrown during docsite CI build
The error
    System.IO.IOException: The process cannot access the file
    '/home/build/RandN/docs/output/assets/css/override.css'
    because it is being used by another process"
was being thrown on occasion when the docsite built on SourceHut, but not locally. The
exception originated in the Resources pipeline. I'm not sure of the exact cause, but
renaming override.css to override.scss (to treat it as a SASS file) fixed the issue, so
I suspect it has to do with the SASS preprocessor.