Add tests covering as much of the ChaCha RNG as possible
Parts of ChaChaIntrinsics cannot be covered easily on platforms supporting
Intrinsics (.NET Core 3.1+), since we intelligently choose whether to run on
AVX2, SSE2, or a software implementation. As far as I know, it's not possible
to set different Enironment Variables for each test natively with XUnit, so
they end up being uncovered.