Fix failure to use optimized bitwise operations instead of backwards-compatible versions
DefineConstants may only be used once per property group, but we used 3 in a single group. This meant that only the last one, BITWISE_FLOAT_CONVERSION was defined. BitOperations.RotateLeft and RotateRight were not used when they should have, and the heap-allocating overload of Decimal.GetBits was used instead of the Span overload.